pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kindex"
version = "0.2.0"
description = "Knowledge index that learns from your conversations"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Jeremy McEntire", email = "jmcentire@users.noreply.github.com" }]
keywords = ["knowledge-graph", "ai", "context", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40"]
dev = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/jmcentire/kindex"
Repository = "https://github.com/jmcentire/kindex"
Issues = "https://github.com/jmcentire/kindex/issues"

[project.scripts]
kin = "kindex.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kindex"]

[tool.pytest.ini_options]
testpaths = ["tests"]
